+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.2.0 (unreleased)
 
 * Bump MSRV to 1.35
+* Drop support for the `rocket_04` Cargo feature (Rocket 0.4 `FromFormValue` / `FromParam`
+  implementations)
 
 # 0.1.9